feat(authorization-scope): align with other enums
