+++ src/main/java/se/michaelthelin/spotify/requests/authorization/authorization_code/AuthorizationCodeUriRequest.java
@@ -2,7 +2,7 @@ package se.michaelthelin.spotify.requests.authorization.authorization_code;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import se.michaelthelin.spotify.SpotifyApi;
-import se.michaelthelin.spotify.enums.AuthScope;
+import se.michaelthelin.spotify.enums.AuthorizationScope;
 import se.michaelthelin.spotify.requests.AbstractRequest;
 
 import java.net.URI;
@@ -140,10 +140,10 @@ public class AuthorizationCodeUriRequest extends AbstractRequest<URI> {
       return setQueryParameter("scope", scope);
     }
 
-    public Builder scope(final AuthScope... scopes) {
+    public Builder scope(final AuthorizationScope... scopes) {
       String finalScopes = "";
 
-      for (AuthScope scope : scopes) {
+      for (AuthorizationScope scope : scopes) {
         finalScopes += (scope.GetScope() + " ");
       }
